requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
PyYAML
wheel
openai==1.3.0
line-bot-sdk==3.1.0
selenium>=4.0.0
beautifulsoup4
pandas
python-dotenv
sentence-transformers>=2.2.2
numpy>=1.21.0
scikit-learn>=1.0.0
torch>=1.9.0
transformers>=4.20.0
joblib>=1.1.0